fix & sort imports. move tooltip to global util
